refactor(storage): 拆分 StoragePane 測試檔並抽共用 fixture
PR-C 前置還債 T4.0 之一（codex 檔案體質 review 在 PR-A / PR-B 各判定一次）。
StoragePane.test.tsx 已膨脹到 1981 行、單一 beforeEach 裡塞了八組 mock
wiring，而 PR-C 還要再往裡面加每列操作、批次選取、0 B 清理與 placeholder。

行為不變，純搬移：
- 依主題拆成 7 個檔案（crud / open / delete-panes / rename-sync / tree /
  dnd / upload），全部移進 storage/__tests__/。測試總數 65 不變。
- 共用 mock 工廠與 spy 收進 __tests__/storage-pane-mocks.tsx；vi.mock 是
  per-file hoist 所以註冊仍留在各檔，但工廠本體只寫一次。
- 共用 fixture 與 beforeEach 重置收進 __tests__/storage-pane-harness.ts。
- 新增 stores/__tests__/editor-buffer-fixture.ts：makeEditorBuffer /
  makeEditorPaneState。之前每個測試都硬編 EditorBuffer 的全部 12 個欄位，
  PR-B 加三個 source* 欄位時就得改遍不相干的 storage 測試；現在 buffer 形狀
  只有這一處。storage-actions.test.ts 的 seedBuffer 一併改用。
