--- spa/src/components/editor/storage/StoragePane.test.tsx
@@ -1,1981 +1,0 @@
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
-import type { Mock } from 'vitest'
-import { useState, type ReactNode } from 'react'
-import { render, screen, fireEvent, waitFor, within, act, createEvent } from '@testing-library/react'
-import { StoragePane } from './StoragePane'
-import { SOFT_MAX_UPLOAD_BYTES } from './storage-actions'
-import { openInAppFile } from '../../../lib/open-in-app-file'
-import { triggerDownload } from '../../../lib/download-file'
-import type { Pane, Tab } from '../../../types/tab'
-import type { FsBackend } from '../../../lib/fs-backend'
-import type { FileEntry, FileStat, FileSource } from '../../../types/fs'
-import type { DragEndEvent } from '@dnd-kit/core'
-
-// dnd-kit is mocked to a lightweight harness: DndContext captures the
-// `onDragEnd` callback so tests can drive a drop synthetically (real pointer
-// drag is impractical in jsdom — the plan endorses a pure handler +
-// onDragEnd-capture smoke). The draggable/droppable hooks become no-ops so rows
-// still render and the existing click/double-click suite keeps passing.
-let mockDragEnd: ((event: DragEndEvent) => void | Promise<void>) | undefined
-vi.mock('@dnd-kit/core', () => ({
-  DndContext: ({ children, onDragEnd }: { children: ReactNode; onDragEnd: (e: DragEndEvent) => void }) => {
-    mockDragEnd = onDragEnd
-    return children
-  },
-  useDraggable: () => ({
-    attributes: {},
-    listeners: {},
-    setNodeRef: () => undefined,
-    transform: null,
-    isDragging: false,
-  }),
-  useDroppable: () => ({ setNodeRef: () => undefined, isOver: false }),
-  PointerSensor: class {},
-  useSensor: () => ({}),
-  useSensors: () => [],
-  closestCenter: () => [],
-}))
-
-// Mock fs-backend: list/write/delete/rename/read are per-test spies. The helper
-// below installs a fresh mock into the registry before each test.
-type MockBackend = {
-  list: Mock
-  write: Mock
-  delete: Mock
-  rename: Mock
-  stat: Mock
-  read: Mock
-  mkdir: Mock
-  mkdirUnique: Mock
-  createUnique: Mock
-  id: 'inapp'
-  label: string
-  available: () => boolean
-}
-
-let mockBackend: MockBackend
-const eventLog: string[] = []
-
-vi.mock('../../../lib/fs-backend', () => ({
-  getFsBackend: () => mockBackend as unknown as FsBackend,
-  registerFsBackend: vi.fn(),
-  // The capability guards (codex H1) narrow a resolved backend; mirror the real
-  // typeof check against the mock backend.
-  supportsCreateUnique: (b: { createUnique?: unknown } | undefined) =>
-    typeof b?.createUnique === 'function',
-  supportsMkdirUnique: (b: { mkdirUnique?: unknown } | undefined) =>
-    typeof b?.mkdirUnique === 'function',
-}))
-
-// Open routing is the StoragePane's collaboration boundary: the pane resolves
-// the workspace id and hands (path, wsId) to `openInAppFile`. The registry
-// kind-dispatch (md→editor / png→image-preview / pdf→pdf-preview) +
-// open-or-focus + insertTab placement are exercised in
-// `lib/open-in-app-file.test.ts`; here we only assert the pane routes through it.
-vi.mock('../../../lib/open-in-app-file', () => ({
-  openInAppFile: vi.fn(() => 'opened-tab'),
-}))
-
-// The Download toolbar button routes through the real `downloadStorageFile`
-// (which reads via the mocked backend) but the final OS download is the shared
-// `triggerDownload` util — stub it so jsdom's missing createObjectURL /
-// navigation never fires, and so we can assert the download was dispatched.
-vi.mock('../../../lib/download-file', () => ({
-  triggerDownload: vi.fn(),
-}))
-
-// Workspace store: StoragePane.resolveWorkspaceId maps the owning tab → its
-// workspace (R2-2 — no activeWorkspaceId guess). The Storage pane lives in
-// `storageTab` (seeded into the tab store below), which this mock maps to ws1;
-// any other / missing tab resolves to null (the "no owning workspace" case).
-vi.mock('../../../features/workspace/store', () => ({
-  useWorkspaceStore: {
-    getState: () => ({
-      activeWorkspaceId: 'ws1',
-      findWorkspaceByTab: (tabId: string) => (tabId === 'storageTab' ? { id: 'ws1' } : null),
-    }),
-  },
-}))
-
-// i18n is a hoisted spy so the upload banner tests (C6) can assert the
-// interpolation PARAMS (name / cap / counts), not just the key. The spy still
-// returns the key with any `{{var}}` placeholders substituted, so the existing
-// `getByText('editor.buffers.…')` assertions (keys carry no placeholders) hold.
-const { tSpy } = vi.hoisted(() => ({
-  tSpy: vi.fn((key: string, params?: Record<string, string | number>): string => {
-    if (!params) return key
-    return key.replace(/\{\{(\w+)\}\}/g, (_, k) => String(params[k] ?? ''))
-  }),
-}))
-
-vi.mock('../../../stores/useI18nStore', () => ({
-  useI18nStore: (selector: (s: { t: (k: string, p?: Record<string, string | number>) => string }) => unknown) =>
-    selector({ t: tSpy }),
-}))
-
-const setPaneContentSpy = vi.fn()
-const setActiveTabSpy = vi.fn()
-const addTabSpy = vi.fn()
-const closePaneSpy = vi.fn()
-const renameEditorPanesSpy = vi.fn()
-
-type TabStoreState = {
-  tabs: Record<string, Tab>
-  tabOrder: string[]
-  activeTabId: string | null
-  setPaneContent: typeof setPaneContentSpy
-  setActiveTab: typeof setActiveTabSpy
-  addTab: typeof addTabSpy
-  closePane: typeof closePaneSpy
-  renameEditorPanes: typeof renameEditorPanesSpy
-}
-
-let tabStoreState: TabStoreState
-
-vi.mock('../../../stores/useTabStore', () => ({
-  useTabStore: {
-    getState: () => tabStoreState,
-  },
-}))
-
-vi.mock('../../RenamePopover', () => ({
-  RenamePopover: ({
-    currentName,
-    onConfirm,
-    onCancel,
-    validateName,
-    error,
-  }: {
-    currentName: string
-    onConfirm: (name: string) => Promise<void>
-    onCancel: () => void
-    validateName?: (name: string, cur: string) => string | undefined
-    error?: string
-  }) => (
-    <RenameHarness
-      currentName={currentName}
-      onConfirm={onConfirm}
-      onCancel={onCancel}
-      validateName={validateName}
-      externalError={error}
-    />
-  ),
-}))
-
-function RenameHarness({
-  currentName,
-  onConfirm,
-  onCancel,
-  validateName,
-  externalError,
-}: {
-  currentName: string
-  onConfirm: (name: string) => Promise<void>
-  onCancel: () => void
-  validateName?: (name: string, cur: string) => string | undefined
-  externalError?: string
-}) {
-  const [value, setValue] = useState(currentName)
-  const validationError = validateName?.(value.trim(), currentName)
-  const error = validationError ?? externalError
-  return (
-    <div data-testid="rename-popover-harness">
-      <input
-        data-testid="rename-input"
-        value={value}
-        onChange={(e: React.ChangeEvent<HTMLInputElement>) => setValue(e.target.value)}
-      />
-      {error && <p data-testid="rename-error">{error}</p>}
-      <button data-testid="rename-confirm" disabled={!!validationError} onClick={() => onConfirm(value.trim())}>
-        confirm
-      </button>
-      <button data-testid="rename-cancel" onClick={onCancel}>
-        cancel
-      </button>
-    </div>
-  )
-}
-
-function makePane(): Pane {
-  return { id: 'bufpane', content: { kind: 'editor-buffers' } }
-}
-
-/**
- * The tab that hosts the Storage pane itself, so `resolveWorkspaceId` finds
- * `bufpane` in a tab and maps it (→ ws1) instead of returning null. Seeded into
- * the default tab store; open tests that reassign `tabStoreState.tabs` spread
- * this back in.
- */
-function makeStorageTab(): Tab {
-  return {
-    id: 'storageTab',
-    pinned: false,
-    locked: false,
-    createdAt: Date.now(),
-    layout: { type: 'leaf', pane: { id: 'bufpane', content: { kind: 'editor-buffers' } } },
-  }
-}
-
-function makeEditorTab(tabId: string, paneId: string, filePath: string, locked = false): Tab {
-  return {
-    id: tabId,
-    pinned: false,
-    locked,
-    createdAt: Date.now(),
-    layout: {
-      type: 'leaf',
-      pane: { id: paneId, content: { kind: 'editor', source: { type: 'inapp' }, filePath } },
-    },
-  }
-}
-
-function makePreviewTab(
-  tabId: string,
-  paneId: string,
-  filePath: string,
-  kind: 'image-preview' | 'pdf-preview',
-  locked = false,
-): Tab {
-  return {
-    id: tabId,
-    pinned: false,
-    locked,
-    createdAt: Date.now(),
-    layout: {
-      type: 'leaf',
-      pane: { id: paneId, content: { kind, source: { type: 'inapp' }, filePath } },
-    },
-  }
-}
-
-function makeNonEditorTab(tabId: string, paneId: string): Tab {
-  return {
-    id: tabId,
-    pinned: false,
-    locked: false,
-    createdAt: Date.now(),
-    layout: { type: 'leaf', pane: { id: paneId, content: { kind: 'dashboard' } } },
-  }
-}
-
-/** Build a path-aware `list` from a flat fixture map (mirrors useStorageTree.test). */
-function pathAwareList(paths: Map<string, { isDir: boolean; size: number }>): Mock {
-  return vi.fn(async (path: string): Promise<FileEntry[]> => {
-    const prefix = path.endsWith('/') ? path : path + '/'
-    const seen = new Map<string, FileEntry>()
-    for (const [p, meta] of paths) {
-      if (!p.startsWith(prefix)) continue
-      const rest = p.slice(prefix.length)
-      if (rest.includes('/')) continue
-      if (!seen.has(rest)) seen.set(rest, { name: rest, isDir: meta.isDir, size: meta.size })
-    }
-    return Array.from(seen.values())
-  })
-}
-
-beforeEach(() => {
-  eventLog.length = 0
-  mockDragEnd = undefined
-  setPaneContentSpy.mockReset()
-  setActiveTabSpy.mockReset()
-  addTabSpy.mockReset()
-  closePaneSpy.mockReset()
-  renameEditorPanesSpy.mockReset()
-  ;(openInAppFile as unknown as Mock).mockReset()
-  ;(openInAppFile as unknown as Mock).mockReturnValue('opened-tab')
-  ;(triggerDownload as unknown as Mock).mockReset()
-  tSpy.mockClear()
-
-  setPaneContentSpy.mockImplementation((tabId: string, paneId: string) => {
-    eventLog.push(`setPaneContent:${tabId}:${paneId}`)
-  })
-  setActiveTabSpy.mockImplementation((tabId: string) => {
-    eventLog.push(`setActiveTab:${tabId}`)
-  })
-  addTabSpy.mockImplementation(() => {
-    eventLog.push('addTab')
-  })
-  closePaneSpy.mockImplementation((tabId: string, paneId: string) => {
-    eventLog.push(`close:${tabId}:${paneId}`)
-  })
-  renameEditorPanesSpy.mockImplementation(
-    (source: FileSource, oldPath: string, newPath: string) => {
-      eventLog.push(`renameEditorPanes:${oldPath}:${newPath}`)
-      const nextTabs: Record<string, Tab> = {}
-      for (const [tabId, tab] of Object.entries(tabStoreState.tabs)) {
-        if (tab.layout.type === 'leaf') {
-          const c = tab.layout.pane.content
-          if (c.kind === 'editor' && c.source.type === source.type && c.filePath === oldPath) {
-            nextTabs[tabId] = {
-              ...tab,
-              layout: {
-                type: 'leaf',
-                pane: { ...tab.layout.pane, content: { ...c, filePath: newPath } },
-              },
-            }
-            continue
-          }
-        }
-        nextTabs[tabId] = tab
-      }
-      tabStoreState.tabs = nextTabs
-    },
-  )
-
-  tabStoreState = {
-    tabs: { storageTab: makeStorageTab() },
-    tabOrder: ['storageTab'],
-    activeTabId: 'storageTab',
-    setPaneContent: setPaneContentSpy,
-    setActiveTab: setActiveTabSpy,
-    addTab: addTabSpy,
-    closePane: closePaneSpy,
-    renameEditorPanes: renameEditorPanesSpy,
-  }
-
-  mockBackend = {
-    id: 'inapp',
-    label: 'In-App Storage',
-    available: () => true,
-    list: vi.fn().mockResolvedValue([] as FileEntry[]),
-    write: vi.fn().mockResolvedValue(undefined),
-    delete: vi.fn().mockImplementation(async (path: string) => {
-      eventLog.push(`delete:${path}`)
-    }),
-    rename: vi.fn().mockResolvedValue(undefined),
-    stat: vi.fn().mockResolvedValue({ size: 0, mtime: 0, isDirectory: false, isFile: true } as FileStat),
-    read: vi.fn().mockResolvedValue(new Uint8Array(0)),
-    mkdir: vi.fn().mockResolvedValue(undefined),
-    mkdirUnique: vi.fn().mockResolvedValue('/buffer/New Folder'),
-    createUnique: vi.fn().mockResolvedValue('/buffer/Untitled.md'),
-  }
-
-  vi.stubGlobal('confirm', vi.fn(() => true))
-})
-
-afterEach(() => {
-  vi.unstubAllGlobals()
-})
-
-describe('StoragePane', () => {
-  it('B2-1: shows empty state when list returns []', async () => {
-    mockBackend.list.mockResolvedValue([] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    await waitFor(() => {
-      expect(screen.getByText('editor.buffers.empty')).toBeTruthy()
-    })
-  })
-
-  it('B2-2: lists entries by name ascending', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'c.md', isDir: false, size: 10 },
-      { name: 'a.md', isDir: false, size: 10 },
-      { name: 'b.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const rows = await screen.findAllByTestId('buffer-row')
-    expect(rows.map((r) => r.getAttribute('data-name'))).toEqual(['a.md', 'b.md', 'c.md'])
-  })
-
-  it('B2-3: delete (single) calls backend.delete with the right path', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'foo.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row) // select
-    fireEvent.click(screen.getByTestId('toolbar-delete'))
-    await waitFor(() => {
-      expect(mockBackend.delete).toHaveBeenCalledWith('/buffer/foo.md', false)
-    })
-  })
-
-  it('B2-4: New calls backend.createUnique(root, Untitled, md) and refreshes', async () => {
-    mockBackend.list.mockResolvedValue([] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-    fireEvent.click(screen.getByTestId('toolbar-new'))
-    await waitFor(() => {
-      expect(mockBackend.createUnique).toHaveBeenCalledTimes(1)
-    })
-    // Eager unified namer (#854): no blind write(), reservation is atomic.
-    expect(mockBackend.createUnique).toHaveBeenCalledWith('/buffer', 'Untitled', 'md')
-    expect(mockBackend.write).not.toHaveBeenCalled()
-    // Refresh: list called a second time once the hook re-reads.
-    await waitFor(() => {
-      expect(mockBackend.list).toHaveBeenCalledTimes(2)
-    })
-  })
-
-  it('B2-5: rename rejects slash via validateName; backend.rename NOT called', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'foo.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    fireEvent.change(input, { target: { value: 'drafts/foo.md' } })
-    expect(screen.getByTestId('rename-error').textContent).toBe('editor.buffers.rename_slash_error')
-    expect((screen.getByTestId('rename-confirm') as HTMLButtonElement).disabled).toBe(true)
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-    expect(mockBackend.rename).not.toHaveBeenCalled()
-  })
-
-  it('B2-6: flat rename calls backend.rename with old and new paths', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'foo.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    mockBackend.stat.mockRejectedValue(new Error('not found'))
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    fireEvent.change(input, { target: { value: 'bar.md' } })
-    expect(screen.queryByTestId('rename-error')).toBeNull()
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-    await waitFor(() => {
-      expect(mockBackend.rename).toHaveBeenCalledWith('/buffer/foo.md', '/buffer/bar.md')
-    })
-  })
-
-  // --- Open via openInAppFile (rewritten smart-open tests B2-7/8/9/15/18) ---
-
-  it('B2-7: toolbar Open routes a .md file through openInAppFile (no hijack)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'target.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    // Editor tabs exist — the old smart-open would have hijacked one of them.
-    tabStoreState.tabs = {
-      storageTab: makeStorageTab(),
-      TA: makeEditorTab('TA', 'P1', '/buffer/other.md'),
-      TB: makeEditorTab('TB', 'P2', '/buffer/another.md'),
-    }
-    tabStoreState.tabOrder = ['TA', 'TB']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row) // select
-    fireEvent.click(screen.getByTestId('toolbar-open'))
-    await waitFor(() => {
-      expect(openInAppFile).toHaveBeenCalledWith('/buffer/target.md', 'ws1')
-    })
-    // No cross-tab hijack: existing editor panes untouched.
-    expect(setPaneContentSpy).not.toHaveBeenCalled()
-    expect(setActiveTabSpy).not.toHaveBeenCalled()
-    expect(addTabSpy).not.toHaveBeenCalled()
-  })
-
-  it('B2-8: opening does not mutate or activate any existing editor tab', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'target.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    tabStoreState.tabs = {
-      storageTab: makeStorageTab(),
-      TA: makeNonEditorTab('TA', 'PA'),
-      TB: makeEditorTab('TB', 'P2', '/buffer/existing.md'),
-    }
-    tabStoreState.tabOrder = ['TA', 'TB']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.doubleClick(row)
-    await waitFor(() => {
-      expect(openInAppFile).toHaveBeenCalledWith('/buffer/target.md', 'ws1')
-    })
-    expect(setPaneContentSpy).not.toHaveBeenCalled()
-    expect(setActiveTabSpy).not.toHaveBeenCalled()
-    expect(addTabSpy).not.toHaveBeenCalled()
-  })
-
-  it('B2-9: double-clicking a .pdf routes through openInAppFile with its full path', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'report.pdf', isDir: false, size: 2048 },
-    ] as FileEntry[])
-    tabStoreState.tabs = { storageTab: makeStorageTab(), TA: makeNonEditorTab('TA', 'PA') }
-    tabStoreState.tabOrder = ['TA']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.doubleClick(row)
-    await waitFor(() => {
-      expect(openInAppFile).toHaveBeenCalledWith('/buffer/report.pdf', 'ws1')
-    })
-    expect(addTabSpy).not.toHaveBeenCalled()
-  })
-
-  it('B2-15: opening a .png never reuses a dirty/non-inapp editor pane', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'pic.png', isDir: false, size: 4096 },
-    ] as FileEntry[])
-    tabStoreState.tabs = {
-      storageTab: makeStorageTab(),
-      TA: makeEditorTab('TA', 'P1', '/buffer/dirty.md', false),
-      TB: {
-        id: 'TB',
-        pinned: false,
-        locked: false,
-        createdAt: Date.now(),
-        layout: {
-          type: 'leaf',
-          pane: {
-            id: 'PD',
-            content: {
-              kind: 'editor',
-              source: { type: 'daemon', hostId: 'remote' },
-              filePath: '/home/user/file.md',
-            },
-          },
-        },
-      },
-    }
-    tabStoreState.tabOrder = ['TA', 'TB']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.doubleClick(row)
-    await waitFor(() => {
-      expect(openInAppFile).toHaveBeenCalledWith('/buffer/pic.png', 'ws1')
-    })
-    expect(setPaneContentSpy).not.toHaveBeenCalled()
-  })
-
-  it('B2-18: double-click opens the row actually clicked (v1.5 G3)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'a.md', isDir: false, size: 10 },
-      { name: 'b.md', isDir: false, size: 10 },
-      { name: 'c.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const rows = await screen.findAllByTestId('buffer-row')
-    // Sorted alphabetically: a, b, c. Double-click b.md (index 1).
-    fireEvent.doubleClick(rows[1])
-    await waitFor(() => {
-      expect(openInAppFile).toHaveBeenCalledWith('/buffer/b.md', 'ws1')
-    })
-    expect(openInAppFile).toHaveBeenCalledTimes(1)
-  })
-
-  it('B2-10: delete closes each open editor pane BEFORE backend.delete', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'x.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    tabStoreState.tabs = {
-      TA: makeEditorTab('TA', 'P1', '/buffer/x.md'),
-      TB: makeEditorTab('TB', 'P2', '/buffer/x.md'),
-    }
-    tabStoreState.tabOrder = ['TA', 'TB']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-delete'))
-    await waitFor(() => {
-      expect(mockBackend.delete).toHaveBeenCalledWith('/buffer/x.md', false)
-    })
-    const firstDeleteIdx = eventLog.findIndex((e) => e.startsWith('delete:'))
-    const closeIdxs = eventLog
-      .map((e, i) => (e.startsWith('close:') ? i : -1))
-      .filter((i) => i !== -1)
-    expect(closeIdxs.length).toBe(2)
-    for (const i of closeIdxs) {
-      expect(i).toBeLessThan(firstDeleteIdx)
-    }
-    expect(closePaneSpy).toHaveBeenCalledWith('TA', 'P1')
-    expect(closePaneSpy).toHaveBeenCalledWith('TB', 'P2')
-  })
-
-  it('B2-11: delete with locked tab is refused outright (v1.4 F2)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'x.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    tabStoreState.tabs = { TA: makeEditorTab('TA', 'P1', '/buffer/x.md', true) }
-    tabStoreState.tabOrder = ['TA']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-delete'))
-    await waitFor(() => {
-      expect(screen.getByText('editor.buffers.delete_locked_refused')).toBeTruthy()
-    })
-    expect(mockBackend.delete).not.toHaveBeenCalled()
-    expect(closePaneSpy).not.toHaveBeenCalled()
-    expect(tabStoreState.tabs.TA).toBeDefined()
-  })
-
-  it('B2-11b: delete closes an open image-preview / pdf-preview pane before backend.delete', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'p.png', isDir: false, size: 10 },
-    ] as FileEntry[])
-    tabStoreState.tabs = {
-      TA: makePreviewTab('TA', 'P1', '/buffer/p.png', 'image-preview'),
-      TB: makePreviewTab('TB', 'P2', '/buffer/p.png', 'pdf-preview'),
-    }
-    tabStoreState.tabOrder = ['TA', 'TB']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-delete'))
-    await waitFor(() => {
-      expect(mockBackend.delete).toHaveBeenCalledWith('/buffer/p.png', false)
-    })
-    const firstDeleteIdx = eventLog.findIndex((e) => e.startsWith('delete:'))
-    const closeIdxs = eventLog
-      .map((e, i) => (e.startsWith('close:') ? i : -1))
-      .filter((i) => i !== -1)
-    expect(closeIdxs.length).toBe(2)
-    for (const i of closeIdxs) {
-      expect(i).toBeLessThan(firstDeleteIdx)
-    }
-    expect(closePaneSpy).toHaveBeenCalledWith('TA', 'P1')
-    expect(closePaneSpy).toHaveBeenCalledWith('TB', 'P2')
-  })
-
-  it('B2-11c: delete refused when a preview pane sits in a locked tab (v1.4 F2)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'p.png', isDir: false, size: 10 },
-    ] as FileEntry[])
-    tabStoreState.tabs = { TA: makePreviewTab('TA', 'P1', '/buffer/p.png', 'image-preview', true) }
-    tabStoreState.tabOrder = ['TA']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-delete'))
-    await waitFor(() => {
-      expect(screen.getByText('editor.buffers.delete_locked_refused')).toBeTruthy()
-    })
-    expect(mockBackend.delete).not.toHaveBeenCalled()
-    expect(closePaneSpy).not.toHaveBeenCalled()
-    expect(tabStoreState.tabs.TA).toBeDefined()
-  })
-
-  it('B2-12: rename aborts with inline error when destination exists (v1.4 F4)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'foo.md', isDir: false, size: 10 },
-      { name: 'bar.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    mockBackend.stat.mockImplementation(async (path: string) => {
-      if (path === '/buffer/bar.md') {
-        return { size: 10, mtime: 0, isDirectory: false, isFile: true } as FileStat
-      }
-      throw new Error('not found')
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    const rows = await screen.findAllByTestId('buffer-row')
-    fireEvent.click(rows[1]) // sorted: bar.md, foo.md → foo.md
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    fireEvent.change(input, { target: { value: 'bar.md' } })
-    expect(screen.queryByTestId('rename-error')).toBeNull()
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-    await waitFor(() => {
-      expect(screen.getByTestId('rename-error').textContent).toBe(
-        'editor.buffers.rename_exists_error',
-      )
-    })
-    expect(mockBackend.rename).not.toHaveBeenCalled()
-  })
-
-  it('B2-13: multi-select delete refused when ANY target is in a locked tab (v1.4 F2)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'x.md', isDir: false, size: 10 },
-      { name: 'y.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    tabStoreState.tabs = {
-      TA: makeEditorTab('TA', 'P1', '/buffer/x.md', false),
-      TB: makeEditorTab('TB', 'P2', '/buffer/y.md', true),
-    }
-    tabStoreState.tabOrder = ['TA', 'TB']
-    tabStoreState.activeTabId = 'TA'
-    render(<StoragePane pane={makePane()} isActive />)
-    const rows = await screen.findAllByTestId('buffer-row')
-    fireEvent.click(rows[0]) // x.md (plain click selects only x.md)
-    // Modifier click adds y.md to the multi-selection (codex B5: plain click no
-    // longer accretes — additive multi-select needs cmd/ctrl/shift).
-    fireEvent.click(rows[1], { ctrlKey: true }) // + y.md
-    fireEvent.click(screen.getByTestId('toolbar-delete'))
-    await waitFor(() => {
-      expect(screen.getByText('editor.buffers.delete_locked_refused')).toBeTruthy()
-    })
-    expect(mockBackend.delete).not.toHaveBeenCalled()
-    expect(closePaneSpy).not.toHaveBeenCalled()
-  })
-
-  it('B2-14: delete confirms with dirty-specific (cancel) and single-specific (OK) messages (v1.4 F5+F6)', async () => {
-    // Sub-case (a): DIRTY buffer → dirty-specific message → cancel → no-op.
-    {
-      mockBackend.list.mockResolvedValue([
-        { name: 'z.md', isDir: false, size: 10 },
-      ] as FileEntry[])
-      tabStoreState.tabs = {
-        TA: makeEditorTab('TA', 'P1', '/buffer/z.md', false),
-      }
-      tabStoreState.tabOrder = ['TA']
-      tabStoreState.activeTabId = 'TA'
-      const editorModule = await import('../../../stores/useEditorStore')
-      editorModule.useEditorStore.setState({
-        buffers: {
-          'inapp:/buffer/z.md': {
-            content: 'dirty',
-            savedContent: '',
-            isDirty: true,
-            lastStat: null,
-            modelId: 'm1',
-            language: 'markdown',
-            languageSource: 'extension',
-            eol: 'lf',
-            encoding: 'utf8',
-            sourceEol: 'lf',
-            sourceTrailingNewline: false,
-            sourceLeadingBlankLines: 0,
-          },
-        },
-        paneStates: {},
-      })
-      const confirmFalse = vi.fn(() => false)
-      vi.stubGlobal('confirm', confirmFalse)
-      const { unmount } = render(<StoragePane pane={makePane()} isActive />)
-      const row = await screen.findByTestId('buffer-row')
-      fireEvent.click(row)
-      fireEvent.click(screen.getByTestId('toolbar-delete'))
-      expect(confirmFalse).toHaveBeenCalledWith('editor.buffers.delete_dirty_confirm')
-      expect(mockBackend.delete).not.toHaveBeenCalled()
-      expect(closePaneSpy).not.toHaveBeenCalled()
-      unmount()
-    }
-
-    // Sub-case (b): CLEAN single-target → single-specific message → OK → delete fires.
-    {
-      mockBackend.list.mockResolvedValue([
-        { name: 'only.md', isDir: false, size: 10 },
-      ] as FileEntry[])
-      tabStoreState.tabs = {}
-      tabStoreState.tabOrder = []
-      tabStoreState.activeTabId = null
-      const editorModule = await import('../../../stores/useEditorStore')
-      editorModule.useEditorStore.setState({ buffers: {}, paneStates: {} })
-      const confirmTrue = vi.fn(() => true)
-      vi.stubGlobal('confirm', confirmTrue)
-      render(<StoragePane pane={makePane()} isActive />)
-      const row = await screen.findByTestId('buffer-row')
-      fireEvent.click(row)
-      fireEvent.click(screen.getByTestId('toolbar-delete'))
-      expect(confirmTrue).toHaveBeenCalledWith('editor.buffers.delete_one_confirm')
-      await waitFor(() => {
-        expect(mockBackend.delete).toHaveBeenCalledWith('/buffer/only.md', false)
-      })
-    }
-  })
-
-  it('B2-16: rename syncs tab layout + editor store (v1.5 G1)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'foo.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    mockBackend.stat.mockRejectedValue(new Error('not found'))
-    tabStoreState.tabs = {
-      T1: makeEditorTab('T1', 'P1', '/buffer/foo.md', false),
-    }
-    tabStoreState.tabOrder = ['T1']
-    tabStoreState.activeTabId = 'T1'
-    const editorModule = await import('../../../stores/useEditorStore')
-    editorModule.useEditorStore.setState({
-      buffers: {
-        'inapp:/buffer/foo.md': {
-          content: 'hello',
-          savedContent: '',
-          isDirty: true,
-          lastStat: null,
-          modelId: 'm-foo',
-          language: 'markdown',
-          languageSource: 'extension',
-          eol: 'lf',
-          encoding: 'utf8',
-          sourceEol: 'lf',
-          sourceTrailingNewline: false,
-          sourceLeadingBlankLines: 0,
-        },
-      },
-      paneStates: {
-        P1: {
-          bufferKey: 'inapp:/buffer/foo.md',
-          editorMode: 'raw',
-          showDiff: false,
-          cursorPosition: { line: 1, column: 1 },
-          monacoViewState: null,
-          tiptapViewState: null,
-        },
-      },
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    fireEvent.change(input, { target: { value: 'bar.md' } })
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-
-    await waitFor(() => {
-      expect(mockBackend.rename).toHaveBeenCalledWith('/buffer/foo.md', '/buffer/bar.md')
-    })
-    await waitFor(() => {
-      const leaf = tabStoreState.tabs.T1.layout
-      expect(leaf.type).toBe('leaf')
-      if (leaf.type !== 'leaf') throw new Error('expected leaf')
-      expect(leaf.pane.content).toMatchObject({
-        kind: 'editor',
-        source: { type: 'inapp' },
-        filePath: '/buffer/bar.md',
-      })
-    })
-    const editorState = editorModule.useEditorStore.getState()
-    expect(editorState.buffers['inapp:/buffer/bar.md']).toBeDefined()
-    expect(editorState.buffers['inapp:/buffer/bar.md']?.isDirty).toBe(true)
-    expect(editorState.buffers['inapp:/buffer/foo.md']).toBeUndefined()
-    expect(editorState.paneStates.P1?.bufferKey).toBe('inapp:/buffer/bar.md')
-    expect(renameEditorPanesSpy).toHaveBeenCalledWith(
-      { type: 'inapp' },
-      '/buffer/foo.md',
-      '/buffer/bar.md',
-    )
-  })
-
-  it('B2-16b: rename across extensions refreshes buffer metadata (v1.5 R6 MED)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'foo.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    mockBackend.stat.mockRejectedValue(new Error('not found'))
-    tabStoreState.tabs = {
-      T1: makeEditorTab('T1', 'P1', '/buffer/foo.md', false),
-    }
-    tabStoreState.tabOrder = ['T1']
-    tabStoreState.activeTabId = 'T1'
-    const editorModule = await import('../../../stores/useEditorStore')
-    editorModule.useEditorStore.setState({
-      buffers: {
-        'inapp:/buffer/foo.md': {
-          content: 'hello',
-          savedContent: '',
-          isDirty: false,
-          lastStat: null,
-          modelId: 'm-foo',
-          language: 'markdown',
-          languageSource: 'extension',
-          eol: 'lf',
-          encoding: 'utf8',
-          sourceEol: 'lf',
-          sourceTrailingNewline: false,
-          sourceLeadingBlankLines: 0,
-        },
-      },
-      paneStates: {
-        P1: {
-          bufferKey: 'inapp:/buffer/foo.md',
-          editorMode: 'raw',
-          showDiff: false,
-          cursorPosition: { line: 1, column: 1 },
-          monacoViewState: null,
-          tiptapViewState: null,
-        },
-      },
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    fireEvent.change(input, { target: { value: 'foo.ts' } })
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-
-    await waitFor(() => {
-      expect(mockBackend.rename).toHaveBeenCalledWith('/buffer/foo.md', '/buffer/foo.ts')
-    })
-    await waitFor(() => {
-      const next = editorModule.useEditorStore.getState().buffers['inapp:/buffer/foo.ts']
-      expect(next).toBeDefined()
-      expect(next?.language).toBe('typescript')
-      expect(next?.languageSource).toBe('extension')
-    })
-  })
-
-  it('B2-16c: rename preserves manual language override (v1.5 R6 MED)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'foo.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    mockBackend.stat.mockRejectedValue(new Error('not found'))
-    tabStoreState.tabs = {
-      T1: makeEditorTab('T1', 'P1', '/buffer/foo.md', false),
-    }
-    tabStoreState.tabOrder = ['T1']
-    tabStoreState.activeTabId = 'T1'
-    const editorModule = await import('../../../stores/useEditorStore')
-    editorModule.useEditorStore.setState({
-      buffers: {
-        'inapp:/buffer/foo.md': {
-          content: 'hello',
-          savedContent: '',
-          isDirty: false,
-          lastStat: null,
-          modelId: 'm-foo',
-          language: 'rust',
-          languageSource: 'manual',
-          eol: 'lf',
-          encoding: 'utf8',
-          sourceEol: 'lf',
-          sourceTrailingNewline: false,
-          sourceLeadingBlankLines: 0,
-        },
-      },
-      paneStates: {
-        P1: {
-          bufferKey: 'inapp:/buffer/foo.md',
-          editorMode: 'raw',
-          showDiff: false,
-          cursorPosition: { line: 1, column: 1 },
-          monacoViewState: null,
-          tiptapViewState: null,
-        },
-      },
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    fireEvent.change(input, { target: { value: 'foo.ts' } })
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-
-    await waitFor(() => {
-      const next = editorModule.useEditorStore.getState().buffers['inapp:/buffer/foo.ts']
-      expect(next).toBeDefined()
-      expect(next?.language).toBe('rust')
-      expect(next?.languageSource).toBe('manual')
-    })
-  })
-
-  it('B2-17: delete cleans editor store for background tabs (v1.5 G2)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'x.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    tabStoreState.tabs = {
-      T1: makeEditorTab('T1', 'P1', '/buffer/x.md', false),
-    }
-    tabStoreState.tabOrder = ['T1']
-    tabStoreState.activeTabId = null // background (not active)
-    const editorModule = await import('../../../stores/useEditorStore')
-    editorModule.useEditorStore.setState({
-      buffers: {
-        'inapp:/buffer/x.md': {
-          content: 'unsaved',
-          savedContent: '',
-          isDirty: true,
-          lastStat: null,
-          modelId: 'm-x',
-          language: 'markdown',
-          languageSource: 'extension',
-          eol: 'lf',
-          encoding: 'utf8',
-          sourceEol: 'lf',
-          sourceTrailingNewline: false,
-          sourceLeadingBlankLines: 0,
-        },
-      },
-      paneStates: {
-        P1: {
-          bufferKey: 'inapp:/buffer/x.md',
-          editorMode: 'raw',
-          showDiff: false,
-          cursorPosition: { line: 1, column: 1 },
-          monacoViewState: null,
-          tiptapViewState: null,
-        },
-      },
-    })
-    vi.stubGlobal('confirm', vi.fn(() => true))
-
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-delete'))
-
-    await waitFor(() => {
-      expect(mockBackend.delete).toHaveBeenCalledWith('/buffer/x.md', false)
-    })
-    const editorState = editorModule.useEditorStore.getState()
-    expect(editorState.buffers['inapp:/buffer/x.md']).toBeUndefined()
-    expect(editorState.paneStates.P1).toBeUndefined()
-    expect(closePaneSpy).toHaveBeenCalledWith('T1', 'P1')
-  })
-
-  it('A2-6: pane keeps rendering when the editor module is disabled', async () => {
-    mockBackend.list.mockResolvedValue([] as FileEntry[])
-    const { rerender } = render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-    const mod = await import('../../../stores/useModuleEnabledStore')
-    mod.useModuleEnabledStore.setState({ enabled: { editor: false }, baseline: null })
-    rerender(<StoragePane pane={makePane()} isActive />)
-    expect(screen.getByText('editor.buffers.empty')).toBeTruthy()
-  })
-
-  // --- Nested tree (Phase 1a full-path identity) ---
-
-  it('N1: renders a nested file under expandable folders a → b', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/a', { isDir: true, size: 0 }],
-        ['/buffer/a/b', { isDir: true, size: 0 }],
-        ['/buffer/a/b/x.md', { isDir: false, size: 12 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    // Folder a is visible; its descendants are hidden until expanded.
-    const folderA = await screen.findByTestId('buffer-row')
-    expect(folderA.getAttribute('data-path')).toBe('/buffer/a')
-    expect(folderA.getAttribute('data-isdir')).toBe('true')
-    expect(screen.queryByText('x.md')).toBeNull()
-
-    // Expand a → b appears; expand b → x.md appears. (Double-click toggles
-    // expand now that single-click selects the folder — T0-1.)
-    fireEvent.doubleClick(folderA)
-    const folderB = await screen.findByText('b')
-    fireEvent.doubleClick(folderB.closest('[data-testid="buffer-row"]')!)
-    const leaf = await screen.findByText('x.md')
-    const leafRow = leaf.closest('[data-testid="buffer-row"]')
-    expect(leafRow?.getAttribute('data-path')).toBe('/buffer/a/b/x.md')
-    // Depth-2 indentation (8 + 2*16 = 40px).
-    expect((leafRow as HTMLElement).style.paddingLeft).toBe('40px')
-  })
-
-  it('N2: same-basename files in different dirs are selected independently (full-path identity)', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/d1', { isDir: true, size: 0 }],
-        ['/buffer/d1/x.md', { isDir: false, size: 3 }],
-        ['/buffer/d2', { isDir: true, size: 0 }],
-        ['/buffer/d2/x.md', { isDir: false, size: 3 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    const d1 = await screen.findByText('d1')
-    const d2 = await screen.findByText('d2')
-    // Double-click expands the folders (single-click now selects — T0-1).
-    fireEvent.doubleClick(d1.closest('[data-testid="buffer-row"]')!)
-    fireEvent.doubleClick(d2.closest('[data-testid="buffer-row"]')!)
-    const leaves = await screen.findAllByText('x.md')
-    const row1 = leaves[0].closest('[data-testid="buffer-row"]') as HTMLElement
-    const row2 = leaves[1].closest('[data-testid="buffer-row"]') as HTMLElement
-    expect(row1.getAttribute('data-path')).toBe('/buffer/d1/x.md')
-    expect(row2.getAttribute('data-path')).toBe('/buffer/d2/x.md')
-    // Select only the first; the second stays unselected.
-    fireEvent.click(row1)
-    expect(row1.getAttribute('aria-selected')).toBe('true')
-    expect(row2.getAttribute('aria-selected')).toBe('false')
-  })
-
-  it('N3: rows render the correct extension/folder icon name', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/dir', { isDir: true, size: 0 }],
-        ['/buffer/dir/note.md', { isDir: false, size: 5 }],
-        ['/buffer/pic.png', { isDir: false, size: 9 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    const rows = await screen.findAllByTestId('buffer-row')
-    const byPath = (p: string) => rows.find((r) => r.getAttribute('data-path') === p)!
-    // Collapsed folder → Folder; png → FilePng.
-    expect(byPath('/buffer/dir').getAttribute('data-icon')).toBe('Folder')
-    expect(byPath('/buffer/pic.png').getAttribute('data-icon')).toBe('FilePng')
-    // Expand the folder (double-click) → it switches to FolderOpen, and the md row resolves.
-    fireEvent.doubleClick(byPath('/buffer/dir'))
-    await waitFor(() => {
-      const updated = screen.getAllByTestId('buffer-row')
-      const dir = updated.find((r) => r.getAttribute('data-path') === '/buffer/dir')!
-      expect(dir.getAttribute('data-icon')).toBe('FolderOpen')
-      const md = updated.find((r) => r.getAttribute('data-path') === '/buffer/dir/note.md')!
-      expect(md.getAttribute('data-icon')).toBe('FileMd')
-    })
-  })
-
-  it('N4: text rows show a word count; binary rows show size only', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'doc.md', isDir: false, size: 17 },
-      { name: 'pic.png', isDir: false, size: 4096 },
-    ] as FileEntry[])
-    mockBackend.read.mockImplementation(async (path: string) => {
-      if (path === '/buffer/doc.md') return new TextEncoder().encode('hello world foo')
-      return new Uint8Array(0)
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    const rows = await screen.findAllByTestId('buffer-row')
-    const md = rows.find((r) => r.getAttribute('data-path') === '/buffer/doc.md')!
-    const png = rows.find((r) => r.getAttribute('data-path') === '/buffer/pic.png')!
-    await waitFor(() => {
-      expect(md.textContent).toContain('3 words')
-    })
-    expect(md.textContent).toContain('17 B')
-    // Binary: size only, never a word count, and read is not attempted.
-    expect(png.textContent).toContain('4096 B')
-    expect(png.textContent).not.toContain('words')
-    expect(mockBackend.read).not.toHaveBeenCalledWith('/buffer/pic.png')
-  })
-
-  it('N4b: a text file over the size cap shows size only and is NOT read (R2-3)', async () => {
-    const big = 300 * 1024 // > WORD_COUNT_MAX_BYTES (256 KiB)
-    mockBackend.list.mockResolvedValue([
-      { name: 'big.md', isDir: false, size: big },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    expect(row.textContent).toContain(`${big} B`)
-    expect(row.textContent).not.toContain('words')
-    expect(mockBackend.read).not.toHaveBeenCalled()
-  })
-
-  it('N4c: an unknown-extension row shows size only and is NOT read (R2-3)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'blob.bin', isDir: false, size: 12 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    expect(row.textContent).toContain('12 B')
-    expect(row.textContent).not.toContain('words')
-    expect(mockBackend.read).not.toHaveBeenCalled()
-  })
-
-  it('N4d: a leading-dot dotfile on the text allowlist still shows a word count (R3)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: '.env', isDir: false, size: 11 },
-    ] as FileEntry[])
-    mockBackend.read.mockImplementation(async (path: string) => {
-      if (path === '/buffer/.env') return new TextEncoder().encode('A=1 B=2 C=3')
-      return new Uint8Array(0)
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    await waitFor(() => expect(row.textContent).toContain('3 words'))
-    expect(mockBackend.read).toHaveBeenCalledWith('/buffer/.env')
-  })
-
-  it('R2-2: passes a null workspace id straight through when the pane has no owning workspace', async () => {
-    // bufpane is in no tab → resolveWorkspaceId returns null (no active guess).
-    tabStoreState.tabs = {}
-    tabStoreState.tabOrder = []
-    tabStoreState.activeTabId = null
-    mockBackend.list.mockResolvedValue([
-      { name: 'x.md', isDir: false, size: 5 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.doubleClick(row)
-    await waitFor(() => {
-      expect(openInAppFile).toHaveBeenCalledWith('/buffer/x.md', null)
-    })
-  })
-
-  it('R2-1: refreshes the tree when openInAppFile aborts (stale/missing entry)', async () => {
-    ;(openInAppFile as unknown as Mock).mockResolvedValue(undefined)
-    mockBackend.list.mockResolvedValue([
-      { name: 'stale.md', isDir: false, size: 5 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    mockBackend.list.mockClear()
-    fireEvent.doubleClick(row)
-    await waitFor(() => {
-      // refresh() bumps the tree nonce → re-lists the root.
-      expect(mockBackend.list).toHaveBeenCalled()
-    })
-  })
-
-  it('N5: rename of a nested leaf file targets its full path (same directory)', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/a', { isDir: true, size: 0 }],
-        ['/buffer/a/b', { isDir: true, size: 0 }],
-        ['/buffer/a/b/x.md', { isDir: false, size: 12 }],
-      ]),
-    )
-    mockBackend.stat.mockRejectedValue(new Error('not found'))
-    render(<StoragePane pane={makePane()} isActive />)
-    // Double-click expands the folders (single-click now selects — T0-1).
-    fireEvent.doubleClick(await screen.findByText('a').then((n) => n.closest('[data-testid="buffer-row"]')!))
-    fireEvent.doubleClick((await screen.findByText('b')).closest('[data-testid="buffer-row"]')!)
-    const leafRow = (await screen.findByText('x.md')).closest('[data-testid="buffer-row"]')!
-    fireEvent.click(leafRow) // select
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    // currentName is the basename of the full path.
-    expect((input as HTMLInputElement).value).toBe('x.md')
-    fireEvent.change(input, { target: { value: 'y.md' } })
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-    await waitFor(() => {
-      expect(mockBackend.rename).toHaveBeenCalledWith('/buffer/a/b/x.md', '/buffer/a/b/y.md')
-    })
-  })
-
-  it('N6: delete of a nested leaf file targets its full path', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/a', { isDir: true, size: 0 }],
-        ['/buffer/a/b', { isDir: true, size: 0 }],
-        ['/buffer/a/b/x.md', { isDir: false, size: 12 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    // Double-click expands the folders (single-click now selects — T0-1).
-    fireEvent.doubleClick((await screen.findByText('a')).closest('[data-testid="buffer-row"]')!)
-    fireEvent.doubleClick((await screen.findByText('b')).closest('[data-testid="buffer-row"]')!)
-    const leafRow = (await screen.findByText('x.md')).closest('[data-testid="buffer-row"]')!
-    fireEvent.click(leafRow)
-    fireEvent.click(screen.getByTestId('toolbar-delete'))
-    await waitFor(() => {
-      expect(mockBackend.delete).toHaveBeenCalledWith('/buffer/a/b/x.md', false)
-    })
-  })
-
-  // --- Folder-selectable tree + target model (Phase 1b T1b-0) ---
-
-  it('T0-1: clicking a folder name selects it (selected style), without expanding or opening', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/dir', { isDir: true, size: 0 }],
-        ['/buffer/dir/x.md', { isDir: false, size: 3 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    const folder = await screen.findByTestId('buffer-row')
-    expect(folder.getAttribute('data-path')).toBe('/buffer/dir')
-    fireEvent.click(folder)
-    // Folder shows the selected style and stays collapsed; no tab opened.
-    expect(folder.getAttribute('aria-selected')).toBe('true')
-    expect(screen.queryByText('x.md')).toBeNull()
-    expect(openInAppFile).not.toHaveBeenCalled()
-  })
-
-  it('T0-1b: the caret toggles expand independently of selection', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/dir', { isDir: true, size: 0 }],
-        ['/buffer/dir/x.md', { isDir: false, size: 3 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByTestId('buffer-row') // wait for the async tree build
-    const rowByPath = (p: string) =>
-      screen.getAllByTestId('buffer-row').find((r) => r.getAttribute('data-path') === p)!
-    fireEvent.click(rowByPath('/buffer/dir')) // select the folder
-    expect(rowByPath('/buffer/dir').getAttribute('aria-selected')).toBe('true')
-    // Expanding via the caret reveals the child but leaves selection intact.
-    fireEvent.click(within(rowByPath('/buffer/dir')).getByTestId('buffer-caret'))
-    await screen.findByText('x.md')
-    expect(rowByPath('/buffer/dir').getAttribute('aria-selected')).toBe('true')
-    expect(rowByPath('/buffer/dir/x.md').getAttribute('aria-selected')).toBe('false')
-  })
-
-  it('T0-2: targetDir derives from selection (none→root, folder→self, file→parent)', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/dir', { isDir: true, size: 0 }],
-        ['/buffer/dir/x.md', { isDir: false, size: 3 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    const region = await screen.findByTestId('storage-tree-region')
-    await screen.findByTestId('buffer-row') // wait for the async tree build
-    const rowByPath = (p: string) =>
-      screen.getAllByTestId('buffer-row').find((r) => r.getAttribute('data-path') === p)!
-    // Nothing selected → storage root.
-    expect(region.getAttribute('data-target-dir')).toBe('/buffer')
-    // Folder selected → the folder itself.
-    fireEvent.click(rowByPath('/buffer/dir'))
-    expect(region.getAttribute('data-target-dir')).toBe('/buffer/dir')
-    // Deselect, expand (caret = no selection), select the nested file → its parent.
-    fireEvent.click(rowByPath('/buffer/dir'))
-    fireEvent.click(within(rowByPath('/buffer/dir')).getByTestId('buffer-caret'))
-    await screen.findByText('x.md')
-    fireEvent.click(rowByPath('/buffer/dir/x.md'))
-    expect(region.getAttribute('data-target-dir')).toBe('/buffer/dir')
-  })
-
-  it('T0-3: double-clicking a folder toggles expand and never opens a tab', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/dir', { isDir: true, size: 0 }],
-        ['/buffer/dir/x.md', { isDir: false, size: 3 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    const folder = await screen.findByTestId('buffer-row')
-    fireEvent.doubleClick(folder)
-    await screen.findByText('x.md') // expanded
-    expect(openInAppFile).not.toHaveBeenCalled()
-  })
-
-  // --- New Folder (mkdir) + New File targetDir wiring (Phase 1b T1b-3) ---
-
-  it('T3-1: New Folder creates a directory (mkdirUnique) that appears in the tree', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>()
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.mkdirUnique = vi.fn(async (dir: string) => {
-      const path = `${dir}/New Folder`
-      paths.set(path, { isDir: true, size: 0 })
-      return path
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-    fireEvent.click(screen.getByTestId('toolbar-new-folder'))
-    await waitFor(() => {
-      // No selection → targets the storage root.
-      expect(mockBackend.mkdirUnique).toHaveBeenCalledWith('/buffer')
-    })
-    const row = await screen.findByTestId('buffer-row')
-    expect(row.getAttribute('data-path')).toBe('/buffer/New Folder')
-    expect(row.getAttribute('data-isdir')).toBe('true')
-    // The new folder is auto-selected.
-    expect(row.getAttribute('aria-selected')).toBe('true')
-  })
-
-  it('T3-2: a New File created with the new folder selected lands inside it (folder accepts children)', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>()
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.mkdirUnique = vi.fn(async (dir: string) => {
-      const path = `${dir}/New Folder`
-      paths.set(path, { isDir: true, size: 0 })
-      return path
-    })
-    mockBackend.createUnique = vi.fn(async (dir: string, base: string, ext: string) => {
-      const path = `${dir}/${base}.${ext}`
-      paths.set(path, { isDir: false, size: 0 })
-      return path
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-    // New Folder → auto-selected + auto-expanded.
-    fireEvent.click(screen.getByTestId('toolbar-new-folder'))
-    await screen.findByTestId('buffer-row') // tree rebuilt → selection resolves to the folder
-    // New File now targets the selected folder (T1b-0 wiring).
-    fireEvent.click(screen.getByTestId('toolbar-new'))
-    await waitFor(() => {
-      expect(mockBackend.createUnique).toHaveBeenCalledWith('/buffer/New Folder', 'Untitled', 'md')
-    })
-    // The child file shows up under the (auto-expanded) folder.
-    await waitFor(() => {
-      const child = screen
-        .getAllByTestId('buffer-row')
-        .find((r) => r.getAttribute('data-path') === '/buffer/New Folder/Untitled.md')
-      expect(child).toBeTruthy()
-    })
-  })
-
-  // --- In-place rename of a folder selection (Phase 1b T1b-4) ---
-
-  it('T4-UI-1: renaming a selected folder routes through backend.rename and re-selects the new path', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>([
-      ['/buffer/dir', { isDir: true, size: 0 }],
-      ['/buffer/dir/x.md', { isDir: false, size: 3 }],
-    ])
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.stat.mockRejectedValue(new Error('not found'))
-    // Re-key the fixture like the real recursive backend.rename (T1b-1).
-    mockBackend.rename = vi.fn(async (from: string, to: string) => {
-      for (const [p, meta] of Array.from(paths)) {
-        if (p === from || p.startsWith(from + '/')) {
-          paths.delete(p)
-          paths.set(to + p.slice(from.length), meta)
-        }
-      }
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    const folder = await screen.findByTestId('buffer-row')
-    fireEvent.click(folder) // select the folder
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    expect((input as HTMLInputElement).value).toBe('dir')
-    fireEvent.change(input, { target: { value: 'docs' } })
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-    await waitFor(() => {
-      expect(mockBackend.rename).toHaveBeenCalledWith('/buffer/dir', '/buffer/docs')
-    })
-    expect(mockBackend.rename).toHaveBeenCalledTimes(1)
-    // Tree rebuilt → the renamed folder is present and re-selected by new path.
-    await waitFor(() => {
-      const row = screen
-        .getAllByTestId('buffer-row')
-        .find((r) => r.getAttribute('data-path') === '/buffer/docs')
-      expect(row).toBeTruthy()
-      expect(row!.getAttribute('aria-selected')).toBe('true')
-    })
-  })
-
-  it('T4-UI-2: folder rename onto an existing name shows the inline exists error (no mutation)', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/a', { isDir: true, size: 0 }],
-        ['/buffer/z', { isDir: true, size: 0 }],
-      ]),
-    )
-    mockBackend.stat.mockImplementation(async (p: string) => {
-      if (p === '/buffer/z') return { size: 0, mtime: 0, isDirectory: true, isFile: false } as FileStat
-      throw new Error('not found')
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    const rows = await screen.findAllByTestId('buffer-row')
-    fireEvent.click(rows.find((r) => r.getAttribute('data-path') === '/buffer/a')!)
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    fireEvent.change(input, { target: { value: 'z' } })
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-    await waitFor(() => {
-      expect(screen.getByTestId('rename-error').textContent).toBe('editor.buffers.rename_exists_error')
-    })
-    expect(mockBackend.rename).not.toHaveBeenCalled()
-  })
-
-  it('T3-3: New Folder does not overwrite an existing folder — it increments the name', async () => {
-    // A "New Folder" already exists (e.g. from a prior session); creating
-    // another must not clobber it (decision 7 — add-reserve increments).
-    const paths = new Map<string, { isDir: boolean; size: number }>([
-      ['/buffer/New Folder', { isDir: true, size: 0 }],
-    ])
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.mkdirUnique = vi.fn(async (dir: string) => {
-      let n = 0
-      let path = `${dir}/New Folder`
-      while (paths.has(path)) {
-        n += 1
-        path = `${dir}/New Folder ${n}`
-      }
-      paths.set(path, { isDir: true, size: 0 })
-      return path
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByTestId('buffer-row')
-    fireEvent.click(screen.getByTestId('toolbar-new-folder'))
-    await waitFor(() => {
-      const names = screen
-        .getAllByTestId('buffer-row')
-        .map((r) => r.getAttribute('data-path'))
-      expect(names).toContain('/buffer/New Folder')
-      expect(names).toContain('/buffer/New Folder 1')
-    })
-  })
-
-  // --- Drag-and-drop move wiring (Phase 1b T1b-6b) ---
-  // The pure resolver `computeMoveFromDragEnd` is unit-tested in
-  // `storage-dnd.test.ts`; the cases below drive the wired `onDragEnd` with a
-  // synthetic drop event whose `over.data.current.targetDir` mirrors what the
-  // row / root droppables publish (codex B1).
-
-  it('T6b-1: dragging a file onto a folder moves it into that folder and refreshes', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>([
-      ['/buffer/dir', { isDir: true, size: 0 }],
-      ['/buffer/a.md', { isDir: false, size: 3 }],
-    ])
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.stat.mockRejectedValue(new Error('not found')) // no collision
-    mockBackend.rename = vi.fn(async (from: string, to: string) => {
-      const meta = paths.get(from)!
-      paths.delete(from)
-      paths.set(to, meta)
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findAllByTestId('buffer-row')
-    expect(mockDragEnd).toBeTypeOf('function') // DndContext mounted + wired
-    await act(async () => {
-      await mockDragEnd!({
-        active: { id: '/buffer/a.md' },
-        over: { id: '/buffer/dir', data: { current: { targetDir: '/buffer/dir' } } },
-      } as unknown as DragEndEvent)
-    })
-    expect(mockBackend.rename).toHaveBeenCalledWith('/buffer/a.md', '/buffer/dir/a.md')
-    // Tree refreshed: the file now lives under the (expandable) folder.
-    await waitFor(() => {
-      const names = screen.getAllByTestId('buffer-row').map((r) => r.getAttribute('data-path'))
-      expect(names).not.toContain('/buffer/a.md')
-    })
-  })
-
-  it('T6b-2: dropping a nested file on the root region moves it to STORAGE_ROOT', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>([
-      ['/buffer/dir', { isDir: true, size: 0 }],
-      ['/buffer/dir/a.md', { isDir: false, size: 3 }],
-    ])
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.stat.mockRejectedValue(new Error('not found'))
-    mockBackend.rename = vi.fn(async (from: string, to: string) => {
-      const meta = paths.get(from)!
-      paths.delete(from)
-      paths.set(to, meta)
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findAllByTestId('buffer-row')
-    await act(async () => {
-      await mockDragEnd!({
-        active: { id: '/buffer/dir/a.md' },
-        over: { id: '/buffer', data: { current: { targetDir: '/buffer' } } },
-      } as unknown as DragEndEvent)
-    })
-    expect(mockBackend.rename).toHaveBeenCalledWith('/buffer/dir/a.md', '/buffer/a.md')
-  })
-
-  it('T6b-3: dropping onto self or own descendant is a no-op (no backend mutation)', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>([
-      ['/buffer/dir', { isDir: true, size: 0 }],
-      ['/buffer/dir/sub', { isDir: true, size: 0 }],
-    ])
-    mockBackend.list = pathAwareList(paths)
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findAllByTestId('buffer-row')
-    // onto self
-    await act(async () => {
-      await mockDragEnd!({
-        active: { id: '/buffer/dir' },
-        over: { id: '/buffer/dir', data: { current: { targetDir: '/buffer/dir' } } },
-      } as unknown as DragEndEvent)
-    })
-    // into own descendant
-    await act(async () => {
-      await mockDragEnd!({
-        active: { id: '/buffer/dir' },
-        over: { id: '/buffer/dir/sub', data: { current: { targetDir: '/buffer/dir/sub' } } },
-      } as unknown as DragEndEvent)
-    })
-    expect(mockBackend.rename).not.toHaveBeenCalled()
-    expect(mockBackend.stat).not.toHaveBeenCalled()
-  })
-
-  it('T6b-3b: dropping onto an existing same-named entry surfaces the inline exists error', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>([
-      ['/buffer/dir', { isDir: true, size: 0 }],
-      ['/buffer/dir/a.md', { isDir: false, size: 3 }],
-      ['/buffer/a.md', { isDir: false, size: 9 }],
-    ])
-    mockBackend.list = pathAwareList(paths)
-    // The target /buffer/dir/a.md already exists.
-    mockBackend.stat.mockImplementation(async (p: string) => {
-      if (paths.has(p)) return { size: 0, mtime: 0, isDirectory: false, isFile: true } as FileStat
-      throw new Error('not found')
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findAllByTestId('buffer-row')
-    await act(async () => {
-      await mockDragEnd!({
-        active: { id: '/buffer/a.md' },
-        over: { id: '/buffer/dir', data: { current: { targetDir: '/buffer/dir' } } },
-      } as unknown as DragEndEvent)
-    })
-    expect(mockBackend.rename).not.toHaveBeenCalled()
-    await waitFor(() => {
-      expect(screen.getByText('editor.buffers.rename_exists_error')).toBeTruthy()
-    })
-  })
-
-  it('T6b-4: click still selects and double-click still opens (no DnD regression)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'a.md', isDir: false, size: 3 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    expect(row.getAttribute('aria-selected')).toBe('true')
-    fireEvent.doubleClick(row)
-    await waitFor(() => {
-      expect(openInAppFile).toHaveBeenCalledWith('/buffer/a.md', 'ws1')
-    })
-  })
-
-  it('B1-UI: dropping a file onto a same-dir sibling is a no-op (targetDir = shared parent)', async () => {
-    // The sibling row publishes /buffer as its targetDir; the move then collapses
-    // to a no-op (targetDir === parentOf(from)) instead of bouncing to root.
-    mockBackend.list.mockResolvedValue([
-      { name: 'a.md', isDir: false, size: 3 },
-      { name: 'b.md', isDir: false, size: 3 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findAllByTestId('buffer-row')
-    await act(async () => {
-      await mockDragEnd!({
-        active: { id: '/buffer/a.md' },
-        over: { id: '/buffer/b.md', data: { current: { targetDir: '/buffer' } } },
-      } as unknown as DragEndEvent)
-    })
-    expect(mockBackend.rename).not.toHaveBeenCalled()
-    expect(mockBackend.stat).not.toHaveBeenCalled()
-  })
-
-  // --- Folder Open guard (codex B3) ---
-
-  it('B3: selecting a folder disables the toolbar Open button (folders are not openable)', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/dir', { isDir: true, size: 0 }],
-        ['/buffer/note.md', { isDir: false, size: 3 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findAllByTestId('buffer-row')
-    const rowByPath = (p: string) =>
-      screen.getAllByTestId('buffer-row').find((r) => r.getAttribute('data-path') === p)!
-    const openBtn = screen.getByTestId('toolbar-open') as HTMLButtonElement
-    // Folder selected → Open disabled, and clicking it never routes to open.
-    fireEvent.click(rowByPath('/buffer/dir'))
-    expect(openBtn.disabled).toBe(true)
-    fireEvent.click(openBtn)
-    expect(openInAppFile).not.toHaveBeenCalled()
-    // A file selection re-enables Open.
-    fireEvent.click(rowByPath('/buffer/note.md'))
-    expect(openBtn.disabled).toBe(false)
-  })
-
-  // --- Download / export a stored file (T1c-2) ---
-
-  it('T2-2a: selecting a file enables Download and clicking dispatches downloadStorageFile', async () => {
-    const bytes = new TextEncoder().encode('hello world')
-    mockBackend.list.mockResolvedValue([
-      { name: 'note.md', isDir: false, size: bytes.byteLength },
-    ] as FileEntry[])
-    mockBackend.read.mockResolvedValue(bytes)
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    const downloadBtn = screen.getByTestId('toolbar-download') as HTMLButtonElement
-    // No selection → disabled.
-    expect(downloadBtn.disabled).toBe(true)
-    fireEvent.click(row) // select the file
-    expect(downloadBtn.disabled).toBe(false)
-    fireEvent.click(downloadBtn)
-    await waitFor(() => {
-      // downloadStorageFile read the file's bytes …
-      expect(mockBackend.read).toHaveBeenCalledWith('/buffer/note.md')
-    })
-    await waitFor(() => {
-      // … and handed them to the shared download util as a Blob named by basename.
-      expect(triggerDownload).toHaveBeenCalledTimes(1)
-    })
-    const [blob, filename] = (triggerDownload as unknown as Mock).mock.calls[0]
-    expect(blob).toBeInstanceOf(Blob)
-    expect(filename).toBe('note.md')
-  })
-
-  it('T2-2b: selecting a folder disables Download and clicking never downloads', async () => {
-    mockBackend.list = pathAwareList(
-      new Map([
-        ['/buffer/dir', { isDir: true, size: 0 }],
-        ['/buffer/note.md', { isDir: false, size: 3 }],
-      ]),
-    )
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findAllByTestId('buffer-row')
-    const rowByPath = (p: string) =>
-      screen.getAllByTestId('buffer-row').find((r) => r.getAttribute('data-path') === p)!
-    const downloadBtn = screen.getByTestId('toolbar-download') as HTMLButtonElement
-    // Folder selected → Download disabled, and a click never reads / downloads.
-    fireEvent.click(rowByPath('/buffer/dir'))
-    expect(downloadBtn.disabled).toBe(true)
-    fireEvent.click(downloadBtn)
-    expect(triggerDownload).not.toHaveBeenCalled()
-    expect(mockBackend.read).not.toHaveBeenCalledWith('/buffer/dir')
-    // A file selection re-enables Download.
-    fireEvent.click(rowByPath('/buffer/note.md'))
-    expect(downloadBtn.disabled).toBe(false)
-  })
-
-  // --- Selection model: plain click never toggles off (codex B5) ---
-
-  it('B5: a second plain click keeps the row selected (no toggle-off)', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'a.md', isDir: false, size: 3 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    expect(row.getAttribute('aria-selected')).toBe('true')
-    // Pre-fix this second plain click toggled the row OFF (the click→click→
-    // dblclick sequence stranded the action target); now it stays selected.
-    fireEvent.click(row)
-    expect(row.getAttribute('aria-selected')).toBe('true')
-  })
-
-  it('B5b: a modifier click toggles the row out of a multi-selection', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'a.md', isDir: false, size: 3 },
-      { name: 'b.md', isDir: false, size: 3 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const rows = await screen.findAllByTestId('buffer-row')
-    fireEvent.click(rows[0]) // a.md
-    fireEvent.click(rows[1], { metaKey: true }) // + b.md
-    expect(rows[0].getAttribute('aria-selected')).toBe('true')
-    expect(rows[1].getAttribute('aria-selected')).toBe('true')
-    // Modifier-click b.md again → toggles it back out.
-    fireEvent.click(rows[1], { metaKey: true })
-    expect(rows[1].getAttribute('aria-selected')).toBe('false')
-    expect(rows[0].getAttribute('aria-selected')).toBe('true')
-  })
-
-  // --- Same-name rename closes the popover without a backend rename (codex B4) ---
-
-  it('B4-UI: confirming a rename with the SAME name closes the popover and never calls backend.rename', async () => {
-    mockBackend.list.mockResolvedValue([
-      { name: 'foo.md', isDir: false, size: 10 },
-    ] as FileEntry[])
-    render(<StoragePane pane={makePane()} isActive />)
-    const row = await screen.findByTestId('buffer-row')
-    fireEvent.click(row)
-    fireEvent.click(screen.getByTestId('toolbar-rename'))
-    const input = await screen.findByTestId('rename-input')
-    expect((input as HTMLInputElement).value).toBe('foo.md')
-    // Confirm with the unchanged name.
-    fireEvent.click(screen.getByTestId('rename-confirm'))
-    await waitFor(() => {
-      expect(screen.queryByTestId('rename-popover-harness')).toBeNull()
-    })
-    expect(mockBackend.rename).not.toHaveBeenCalled()
-    expect(mockBackend.stat).not.toHaveBeenCalled()
-  })
-
-  // --- D2: rapid double new-file converges on the atomic namer (no shared key) ---
-
-  it('D2: rapid double New File routes through the atomic createUnique namer with no shared key', async () => {
-    let n = 0
-    mockBackend.list.mockResolvedValue([] as FileEntry[])
-    // Atomic reservation: each call hands back a distinct path (mirrors the IDB
-    // `add` serialization point) — never a Date.now()-style shared key.
-    mockBackend.createUnique = vi.fn(async () => {
-      n += 1
-      return n === 1 ? '/buffer/Untitled.md' : `/buffer/Untitled-${n - 1}.md`
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-    const newBtn = screen.getByTestId('toolbar-new')
-    // Two rapid clicks. The busy-disable guard may serialize them to a single
-    // reservation, but whatever fires goes through the atomic namer — no blind
-    // write, and any reservations are distinct (no shared key, #854).
-    fireEvent.click(newBtn)
-    fireEvent.click(newBtn)
-    await waitFor(() => {
-      expect(mockBackend.createUnique).toHaveBeenCalled()
-    })
-    expect(mockBackend.write).not.toHaveBeenCalled()
-    const reserved = await Promise.all(
-      (mockBackend.createUnique as Mock).mock.results.map((r) => r.value),
-    )
-    // No two reservations collide on the same key.
-    expect(new Set(reserved).size).toBe(reserved.length)
-  })
-})
-
-// --- T1c-1: upload (file picker + native OS-file drop) -----------------------
-//
-// `openInAppFile` is mocked here (its kind dispatch — png→image-preview /
-// pdf→pdf-preview / docx→download — is proven in lib/open-in-app-file.test.ts).
-// These cases assert the StoragePane wiring: the picker / native drop ingests
-// files via the atomic `createUnique(dir, base, ext, bytes)`, the file appears
-// in the tree after refresh, opening routes through `openInAppFile`, and a
-// no-files drop never touches upload (dnd-kit node move untouched).
-
-describe('StoragePane — upload via picker + native OS-file drop (T1c-1)', () => {
-  /** A `createUnique` that records bytes and materializes the path in `paths`. */
-  function wireUploadBackend() {
-    const paths = new Map<string, { isDir: boolean; size: number }>()
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.createUnique = vi.fn(async (dir: string, base: string, ext: string) => {
-      const fileName = ext === '' ? base : `${base}.${ext}`
-      const path = `${dir}/${fileName}`
-      paths.set(path, { isDir: false, size: 1 })
-      return path
-    })
-    return paths
-  }
-
-  async function uploadViaPicker(files: File[]) {
-    const input = screen.getByTestId('upload-input') as HTMLInputElement
-    await act(async () => {
-      fireEvent.change(input, { target: { files } })
-    })
-  }
-
-  it('T1-1: uploading an image via the picker stores its bytes (png ext), shows it in the tree, and opening routes through openInAppFile', async () => {
-    wireUploadBackend()
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    await uploadViaPicker([new File([new Uint8Array([1, 2, 3])], 'photo.png', { type: 'image/png' })])
-
-    await waitFor(() => expect(mockBackend.createUnique).toHaveBeenCalled())
-    const [dir, base, ext, bytes] = (mockBackend.createUnique as Mock).mock.calls[0]
-    expect(dir).toBe('/buffer')
-    expect(base).toBe('photo')
-    expect(ext).toBe('png')
-    expect(Array.from(bytes as Uint8Array)).toEqual([1, 2, 3])
-
-    const row = await screen.findByTestId('buffer-row')
-    expect(row.getAttribute('data-path')).toBe('/buffer/photo.png')
-    fireEvent.doubleClick(row)
-    await waitFor(() => expect(openInAppFile).toHaveBeenCalledWith('/buffer/photo.png', 'ws1'))
-  })
-
-  it('T1-1b: uploading a .pdf stores it and opening routes through openInAppFile (→ pdf-preview)', async () => {
-    wireUploadBackend()
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    await uploadViaPicker([new File([new Uint8Array([8])], 'doc.pdf', { type: 'application/pdf' })])
-
-    await waitFor(() => expect(mockBackend.createUnique).toHaveBeenCalled())
-    expect((mockBackend.createUnique as Mock).mock.calls[0][2]).toBe('pdf')
-    const row = await screen.findByTestId('buffer-row')
-    expect(row.getAttribute('data-path')).toBe('/buffer/doc.pdf')
-    fireEvent.doubleClick(row)
-    await waitFor(() => expect(openInAppFile).toHaveBeenCalledWith('/buffer/doc.pdf', 'ws1'))
-  })
-
-  it('T1-2: uploading a .docx stores it and opening routes through openInAppFile (→ download)', async () => {
-    wireUploadBackend()
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    await uploadViaPicker([new File([new Uint8Array([5])], 'sheet.docx')])
-
-    await waitFor(() => expect(mockBackend.createUnique).toHaveBeenCalled())
-    expect((mockBackend.createUnique as Mock).mock.calls[0][2]).toBe('docx')
-    const row = await screen.findByTestId('buffer-row')
-    expect(row.getAttribute('data-path')).toBe('/buffer/sheet.docx')
-    fireEvent.doubleClick(row)
-    await waitFor(() => expect(openInAppFile).toHaveBeenCalledWith('/buffer/sheet.docx', 'ws1'))
-  })
-
-  it('T1-4: a native drop carrying dataTransfer.files ingests them', async () => {
-    wireUploadBackend()
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    const region = screen.getByTestId('storage-tree-region')
-    await act(async () => {
-      fireEvent.drop(region, {
-        dataTransfer: { files: [new File([new Uint8Array([9])], 'dropped.txt')], types: ['Files'] },
-      })
-    })
-
-    await waitFor(() => expect(mockBackend.createUnique).toHaveBeenCalled())
-    const [dir, base, ext] = (mockBackend.createUnique as Mock).mock.calls[0]
-    expect(dir).toBe('/buffer')
-    expect(base).toBe('dropped')
-    expect(ext).toBe('txt')
-    const row = await screen.findByTestId('buffer-row')
-    expect(row.getAttribute('data-path')).toBe('/buffer/dropped.txt')
-  })
-
-  it('T1-4: a native drop with NO files is ignored (dnd-kit node move untouched)', async () => {
-    wireUploadBackend()
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    const region = screen.getByTestId('storage-tree-region')
-    fireEvent.drop(region, { dataTransfer: { files: [], types: [] } })
-    // Let any stray microtask settle.
-    await act(async () => {
-      await Promise.resolve()
-    })
-    expect(mockBackend.createUnique).not.toHaveBeenCalled()
-  })
-
-  it('C3: dropping an OS FOLDER (types:[Files] but files:[]) is preventDefaulted and not ingested', async () => {
-    wireUploadBackend()
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    const region = screen.getByTestId('storage-tree-region')
-    // An OS folder drag reports the `Files` type but yields an EMPTY files list.
-    // The handler must claim it (preventDefault) so it never falls back to the
-    // browser default — yet must NOT try to ingest a zero-file drop.
-    const dropEvent = createEvent.drop(region, {
-      dataTransfer: { files: [], types: ['Files'] },
-    })
-    await act(async () => {
-      fireEvent(region, dropEvent)
-    })
-    expect(dropEvent.defaultPrevented).toBe(true)
-    expect(mockBackend.createUnique).not.toHaveBeenCalled()
-  })
-
-  it('T1-6: a partial failure surfaces an inline banner naming the first failed file', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>()
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.createUnique = vi.fn(async (dir: string, base: string, ext: string) => {
-      if (base === 'bad') throw new Error('boom')
-      const fileName = ext === '' ? base : `${base}.${ext}`
-      const path = `${dir}/${fileName}`
-      paths.set(path, { isDir: false, size: 1 })
-      return path
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    await uploadViaPicker([
-      new File([new Uint8Array([1])], 'ok.txt'),
-      new File([new Uint8Array([2])], 'bad.bin'),
-    ])
-
-    // Banner reflects partial success. C6: assert the i18n call carries the
-    // right interpolation params (uploaded/failed counts + the first failed
-    // name), not just the key.
-    await waitFor(() => expect(screen.getByText('editor.buffers.upload_partial')).toBeTruthy())
-    expect(tSpy).toHaveBeenCalledWith('editor.buffers.upload_partial', {
-      uploaded: 1,
-      failed: 1,
-      name: 'bad.bin',
-    })
-    const calls = (mockBackend.createUnique as Mock).mock.calls.map((c) => c[1])
-    expect(calls).toEqual(['ok', 'bad'])
-  })
-
-  // --- T1c-4: size cap warning + quota error banner -------------------------
-
-  /** A File whose `size` is forced past the cap WITHOUT allocating the bytes. */
-  function oversizedFile(name: string, size: number): File {
-    const file = new File([new Uint8Array([0])], name)
-    Object.defineProperty(file, 'size', { value: size })
-    return file
-  }
-
-  it('T4-1: an over-cap file shows the amber too-large WARNING and never writes', async () => {
-    wireUploadBackend()
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    await uploadViaPicker([oversizedFile('huge.bin', SOFT_MAX_UPLOAD_BYTES + 1)])
-
-    const warning = await screen.findByTestId('storage-warning')
-    expect(warning.textContent).toBe('editor.buffers.upload_too_large')
-    expect(warning.className).toContain('text-amber-400')
-    // C6: the warning interpolates the file name + the cap in MB (25 = 25 MiB).
-    expect(tSpy).toHaveBeenCalledWith('editor.buffers.upload_too_large', {
-      name: 'huge.bin',
-      cap: 25,
-    })
-    expect(mockBackend.createUnique).not.toHaveBeenCalled()
-  })
-
-  it('T4-2: a quota write failure shows the red quota ERROR banner (not silent)', async () => {
-    const paths = new Map<string, { isDir: boolean; size: number }>()
-    mockBackend.list = pathAwareList(paths)
-    mockBackend.createUnique = vi.fn(async () => {
-      throw new DOMException('full', 'QuotaExceededError')
-    })
-    render(<StoragePane pane={makePane()} isActive />)
-    await screen.findByText('editor.buffers.empty')
-
-    await uploadViaPicker([new File([new Uint8Array([1])], 'x.txt')])
-
-    const banner = await screen.findByText('editor.buffers.upload_quota')
-    expect(banner.className).toContain('text-red-400')
-    // C6: the quota error names the file that could not be saved.
-    expect(tSpy).toHaveBeenCalledWith('editor.buffers.upload_quota', { name: 'x.txt' })
-    expect(screen.queryByTestId('storage-warning')).toBeNull()
-  })
-})
